% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcu_med.R
\name{bind_cut_cate}
\alias{bind_cut_cate}
\title{Extract the statistical summary generated by \code{calcu} for classified data}
\usage{
bind_cut_cate(
  gt_table_body,
  group_list,
  merge_subgroup_name = NULL,
  ppop = NULL
)
}
\arguments{
\item{gt_table_body}{result of \code{calcu} calculation.}

\item{group_list}{need to provide a specified column name, which is usually
a basic diagnosis or drug list.}

\item{merge_subgroup_name}{default is \code{NULL}.Need to provide the specified column name, which
is generally the collective term of basic diagnosis or basic drugs.If this parameter provides a
non-NULL character, then \verb{group_ list} will be regarded as a subset of the diagnosis or drug,
and its corresponding percentage will also be used as the denominator.}

\item{ppop}{default is \code{NULL}. \code{gt_table_body} come from a subset of the
total population? If you need to calculate the statistics of the subset in the
total population, please provide a total population.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A structured data set.A structured dataset. Contains the number and proportion
of subsets after the classification of classified variables.
}\if{html}{\out{</div>}}
}
\description{
Extract the statistical summary generated by \code{calcu} for classified data
}
\note{
When the parameter \code{ppop} is provided, only the parameter \code{ppop} is used by default to
calculate the \code{group_list}. So please redesign the running rules, or use \verb{? bind_cut_ cate} to
view the internal build logic of the function to redesign the code.
}
\examples{
\dontrun{
data(data_med)
diags = data_med$diag
aim_select = c("patient_id","visit_id","diag_union")

diag_wider = wider_clean(diags,aim_select)

diag_wider_calcu = diag_wider \%>\% select(!"patient_id") \%>\%
  dplyr::mutate("vascular disease" =1) \%>\% calcu(.)

group_list = c("Dyslipidemia","high blood pressure","myocardial infarction")
merge_subgroup_name = "vascular disease"
#'
bind_cut_cate(diag_wider_calcu,group_list,merge_subgroup_name)
bind_cut_cate(diag_wider_calcu,group_list,ppop=1000)
}
}
