% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcu_age.R
\name{calcu_age}
\alias{calcu_age}
\title{Calculate the age of the patient}
\usage{
calcu_age(data, birth_date, refDate = Sys.Date(), unit = "year")
}
\arguments{
\item{data}{a dataframe contain  birth date,and alse may be exist index date columns.}

\item{birth_date}{need a colname contain time sequence;
maybe you need to transfer using \code{\link[=as.Date]{as.Date()}}.}

\item{refDate}{Default is in today.
\code{RefDate} needs to use \verb{Index Date} in most statistical scenarios.}

\item{unit}{Default is in years.
Argument 'unit' must be one of 'year', 'month', 'week', or 'day'.}
}
\value{
add a numeric type column called \code{age}
}
\description{
\code{RefDate} needs to use \verb{Index Date} in most statistical scenarios
}
\examples{
data(data_med)
## Calculate the age of the patient based on the present time
data_med$pat \%>\% calcu_age(.,birthdate)
## Calculate the age of the patient based on the index date
data_med$pat \%>\% calcu_age(.,birthdate,index_date)

}
