% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cutcolumn.R
\name{cut_label}
\alias{cut_label}
\title{Provide grouping tags for \code{\link{cut}}}
\usage{
cut_label(breaks_list, openr = FALSE, digits = 0)
}
\arguments{
\item{breaks_list}{data split points for grouping,such as
\code{c(18,30,40,50,60,70,Inf)}}

\item{openr}{default \code{openright= FALSE},the data interval is in the
form of left opening and right closing,such as \verb{[1,2)}.When \code{openright= TRUE},
the data interval is in the form of left closing and right opening.}

\item{digits}{defines the number of decimal places for the return label.}
}
\value{
a grouping tags
}
\description{
Provide grouping tags for \code{\link{cut}}
}
\examples{
breaks_age <- c(18,30,40,50,60,70,Inf)
cut_label(breaks_age)
cut_label(breaks_age,2)

}
