% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_med.R
\docType{data}
\name{data_med}
\alias{data_med}
\title{Virtual medical data set built into \code{pmed}}
\format{
A list of dataframes
}
\usage{
data_med
}
\description{
Returns a data set of 50 patients, including a doctor's table,
prescription table, diagnosis table, and biochemical test table.
}
\keyword{datasets}
