% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkg.R
\name{load_pkg}
\alias{load_pkg}
\title{packages of medical statistics}
\usage{
load_pkg(data)
}
\arguments{
\item{data}{any param is ok!}
}
\value{
return a '?pmed::load_pkg'
}
\description{
List of commonly used R packages of medical statistics
}
\details{
commonly：pacman::p_load(purrr,tidyverse,scales,dplyr,flextable,gt,gtsummary,
rstatix,forcats,janitor,table1,tableone,arrow,
data.table,datacleanr,stringr,tidyr,lubridate,openxlsx,gt)

for exploratory:
rstatix::get_summary_stats(linelist);
dplur::ff_glimpse();
skimr::skim(linelist);
datacleanr::dcr_app(iris);
summarytools::descr(iris);

for descriptive table:
pacman::p_load(knitr,flextable,gt,gtsummary,tableone)

for reactive table:
pacman::p_load(DT)

for speed up data analysis:
pacman::p_load(data.table,arrow);

for model:
pacman::p_load(finalfit,broom)

for plot:
pacman::p_load(ggplot2,plotly)
}
