% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd_pop_perc.R
\name{mean_sd}
\alias{mean_sd}
\title{Calculate \code{mean} and \code{sd} of a given variable}
\usage{
mean_sd(data, aim_var, ifunion = FALSE, digits = 2)
}
\arguments{
\item{data}{a dataframe}

\item{aim_var}{user-specified variables, using non-standard input.}

\item{ifunion}{default \code{FALSE}.when \code{ifunite = TRUE}，return a character
contain \code{mean(sd)}.}

\item{digits}{indicating the number of decimal places (round) or significant digits (signif)
to be used. For round, negative values are allowed.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{`mean_sd` return a data.frame which include two colunms.if `ifunion = true`
retunr a character ,sucn as "mean(sd)".
}\if{html}{\out{</div>}}
}
\description{
Calculate \code{mean} and \code{sd} of a given variable
}
\examples{
\dontrun{
data(data_med)
pat_table = data_med$pat \%>\% calcu_age(.,birthdate)
mead_sd(pat_table,age)
pat \%>\% mean_sd(age)
pat \%>\% mean_sd(age,ifunion =TRUE)
}

}
