% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cutcolumn.R
\name{mutate_class}
\alias{mutate_class}
\title{Convert one continuous variables into grouped variables}
\usage{
mutate_class(data, breaks_list, sp_column, openr = FALSE, digits = 0)
}
\arguments{
\item{data}{a dataframe contain continuous type data column.}

\item{breaks_list}{data split points for grouping such as
\code{c(18,30,40,50,60,70,Inf)}.}

\item{sp_column}{the specified column name adapted to breaks_list,using a
non-standard normal form.}

\item{openr}{default \code{openright= FALSE},the data interval is in the
form of left opening and right closing,such as \verb{[1,2)}.When \code{openright= TRUE},
the data interval is in the form of left closing and right opening.}

\item{digits}{defines the number of decimal places for the return label.}
}
\value{
add a user-specified classified data column to the dataframe.
}
\description{
Generate a new table containing classified columns from continuous variables according to the
grouping interval list provided by the user
}
\details{
It is suitable for the interval segmentation of continuous data such as
patient age and some biochemical indexes.
}
\examples{
\dontrun{
data(data_med)
pat = data_med$pat \%>\% calcu_age(.,birthdate)

breaks_age <- c(18,30,40,50,60,70,Inf)
mutate_class(pat,breaks_age,age,digits=2)
}

}
