% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe.R
\name{pe}
\alias{pe}
\title{Count the number of patients in the table.}
\usage{
pe(data, raw_patient_id)
}
\arguments{
\item{data}{a table contian a  "patient_id" column name or a similar column}

\item{raw_patient_id}{raw patient_id ,use for count ;Non-standard input}
}
\value{
Total number of people after duplicate removal
}
\description{
count patient_id;
}
\details{
need to change the patient id column in the table to "paitent_id"
}
\examples{
datatable1 = data.frame(
patient_id = paste0("patent",1:100),
visit_id  = paste0("visit",1:100))

datatable2 = data.frame(
 patient = paste0("patent",1:100),
 visit_id  = paste0("visit",1:100))

pe(datatable1)

pe(datatable2,patient)
}
