% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tr}
\alias{tr}
\alias{tr_nst}
\title{Multivariable type conversion}
\usage{
tr(data, sp_column, funss)

tr_nst(data, sp_column, funss)
}
\arguments{
\item{data}{a dataframe contain  multiple variable types.}

\item{sp_column}{The column name used to convert the variable type,
use non-standard input in \code{\link[=tr_nst]{tr_nst()}}.}

\item{funss}{\code{"num"} : \code{\link[base:numeric]{base::as.numeric()}};
\code{"dat"}:  \code{\link[base:as.Date]{base::as.Date()}};
\code{"cha"}:  \code{\link[base:character]{base::as.character()}};
\code{"fac"}:  \code{\link[base:factor]{base::as.factor()}}.
use non-standard input in \code{\link[=tr_nst]{tr_nst()}}.}
}
\value{
Convert according to the specified type.
}
\description{
Convert time, characters, numbers and factors;
Support tidy system;
}
\details{
\code{\link[=tr]{tr()}} Use the character definition variable method to convert one or more variables;
\code{\link[=tr_nst]{tr_nst()}} use non-standard variable definition methods to convert a variable type.
}
\examples{
data(data_med)
data_med$pat \%>\% tr(.,c("patient_id","visit_id"),"cha")
data_med$pat \%>\% tr_nst(.,patient_id,num)
}
