% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_med_mulp.R
\name{Virtual_med_one}
\alias{Virtual_med_one}
\title{calculate the virtual dataset of a patient}
\usage{
Virtual_med_one(
  birth_start_date,
  birth_end_date,
  study_start_date,
  study_end_date,
  average_year_visit_n
)
}
\arguments{
\item{birth_start_date}{Minimum expected birth date of the patient.}

\item{birth_end_date}{Maximum expected birth date of the patient.}

\item{study_start_date}{The starting time of the baseline period specified in the study.}

\item{study_end_date}{The end time of the follow-up period specified in the study}

\item{average_year_visit_n}{The average number of visits by patients each year is 6 by default.
The data are from the statistical records of the Chinese people.}
}
\value{
Returns a data set of a patients, including a doctor's table,
prescription table, diagnosis table, and biochemical test table.
}
\description{
calculate the virtual dataset of a patient
}
\examples{
Virtual_med_one(
 birth_start_date = "1980-01-01",birth_end_date ="2000-12-31" ,
   study_start_date = "2014-01-01",study_end_date ="2020-12-31",
   average_year_visit_n = 6)

}
